Show scancodes in debugfs (#4)
* Changed US keys to keycode, also added shift mask

* added the module param

* Tweaked style to fit standard

* Used removed all code formatting errors via checkpatch.pl

* Add feature log raw keycodes

Add a module parameter to set if the module
will log the US Keyboard characters
or the raw keycode values. In raw mode the
shift mask is also added to beginning of each line.

This is done by adding the integer flag "codes" as
a module parameter. If the flag is 0 (default) the behavior
is unchanged from the original version. Otherwise
the shift mask and keycode are converted to their
base 10 string representations and written to the
"keys" log file in the format "mask_space_keycode".

* Removed unecessary example

* Add fun c -> str (still needs code format)

Was on a different machine, will reformat when I have the check script